Update plugin for Blender 2.90 and add support for modifiers
This PR adds a menu item for `V-HACD` to the `Object` menu in the
3D view of blender, as starting with version 2.90, Blender won't show
operators in the operator search unless they've got a corresponding
menu entry.

This also adds a checkbox to the operator settings that allows the user
to generate geometry based on the mesh after applying all moderators,
rather than its shape in edit mode. This is useful when mesh geometry is
changed significantly by deformation modifiers.